requirements.txt
# Core Data Processing
pandas>=1.5.3
numpy>=1.24.3
sqlalchemy>=2.0.15

# Machine Learning
scikit-learn>=1.2.2
scipy>=1.10.1

# Explainability
shap>=0.41.0

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1

# Dashboard
streamlit>=1.22.0
streamlit-aggrid>=0.3.4

# Data Transformation
dbt-core>=1.5.0
dbt-duckdb>=1.5.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.6

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
openpyxl>=3.1.2

# Type Hints
mypy>=1.3.0
types-requests>=2.31.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
